% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixes-extensions.R
\name{args}
\alias{args}
\title{Rewrite of args(), which also returns something workable for all primitives}
\usage{
args(name)
}
\arguments{
\item{name}{A function, or a character string with the name of a function (which is found using the scope of the caller).}
}
\value{
Identical as that of \code{\link[base]{args}}, except when called with a primitive, and args() returns NULL.
In that case, an empty function
}
\description{
The function \code{\link[base]{args}} may return NULL for some primitives, e.g. `[`.
This function checks for that, and in that case returns the most general function possible:
\code{function(...) NULL}
Otherwise the return is identical
}
