% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{duplicated}
\alias{duplicated}
\title{Determine duplicate elements, possibly in both directions}
\usage{
duplicated(x, incomparables = FALSE, fromLast = FALSE, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}

\item{incomparables}{a vector of values that cannot be compared. FALSE is a special value, meaning that all values can be compared, and may be the only value accepted for methods other than the default. It will be coerced internally to the same type as x.}

\item{fromLast}{logical indicating if duplication should be considered from the reverse side (TRUE), the start (FALSE), or from both sides (NA)}

\item{...}{arguments for particular methods}
}
\description{
Determine duplicate elements, possibly in both directions
}
\details{
In the case of \code{fromLast} either \code{TRUE} or \code{FALSE}, \code{\link[base:duplicated]{base::duplicated}} is called directly.
Only when it's \code{NA}, comparison is done from both sides
}
