% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{extractComments}
\alias{extractComments}
\title{Function that could be used as an alternative for other automatic documentation systems.}
\usage{
extractComments(FileName, max = 9, fromLine = 1, ToLine = -1,
  tab = "\\t")
}
\description{
Function that could be used as an alternative for other automatic documentation systems.
}
