% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixes-extensions.R
\name{formalArgs}
\alias{formalArgs}
\title{Bugfix for formalArgs(def)}
\usage{
formalArgs(def)
}
\arguments{
\item{def}{Either a function, or a character naming a function}
}
\description{
In the methods-package (loaded by default), formalArgs(def) uses the search path from namespace:methods, which can lead to confusing behaviour if you overwrite something,
especially as formals(def) DOES use the environment of the caller.
}
\details{
Example of different behaviour:\cr
\code{by <- function(a, b ,c) {"Bye-bye!"}
formalArgs(by) # As expected, by does not have any quotes
names(formals('by')) # Also works as expected
methods::formalArgs('by') # Bug: by is found in the base-package
formalArgs('by') # Result as expected
}
}
