% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{format.Date}
\alias{format.Date}
\title{Alternative for format.Date}
\usage{
\method{format}{Date}(x, ...)
}
\arguments{
\item{x}{Date to format}

\item{...}{Other arguments passed on to format.POSIXlt}
}
\description{
On the R-devel mailinglist, it was noted that as.Date(Inf, origin='1970-01-01') is stored as a valid Date-object,
and is.na() returns FALSE, but when printing this object is shows 'NA", which is confusing.
It turns out this is because when formatting a Date-object it is converted to a POSIXlt-object, which fails for Inf, as well as other out-of-range values.
Therefore this function default to a numerical value if the date is outside the range 1-1-1 up till 9999-12-31, with a warning
}
