% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{libinstandload}
\alias{libinstandload}
\title{Library Auto-installation and loading}
\usage{
libinstandload(..., order = FALSE, verbose = FALSE, extras = TRUE)
}
\arguments{
\item{...}{list of packages to install}

\item{order}{logical specifying whether packages should be loaded in order. If TRUE, first argument is loaded first (so comes last in search())
if FALSE, R tries to load packages first, then installs any misses, so new packages are on top. FALSE is slightly faster}

\item{verbose}{logical, whether to print messages from require and install.package}

\item{extras}{logical. Some packages need extra initialization, e.g. extrafonts needs to install fonts. Should extra initializations be done on installation?\cr
Note that initiliazation scripts are added on an ad-hoc basis, and is not complete. In this first version. it is only implemented for 'extrafont'}
}
\description{
One combined function to load (multiple) packages, and automatically install them if needed
}
