% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{not-in}
\alias{not-in}
\alias{\%!in\%}
\title{Value not matching}
\usage{
x \%!in\% y
}
\arguments{
\item{x, y}{Used as in !(x \%in\% y)}
}
\description{
Simple function, element x is not in y.\cr
\code{x \%!in\% y} is the same as \code{!x \link[base]{\%in\%} y}\cr
Just implemented because I've found myself having to go back too often.
}
\seealso{
\code{\link[base]{+}}, \code{\link[base]{-}}, \code{\link[base]{/}}, \code{\link[base]{\%in\%}}, \code{\link[base]{?}}, \code{\link[base]{::}}, \code{\link[base]{:}}, \code{\link[base]{!}}
}
