% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixes-extensions.R
\name{vmapply}
\alias{vmapply}
\title{Combination of vapply and mapply}
\usage{
vmapply(FUN, FUN.VALUE, ..., MoreArgs = NULL, USE.NAMES = TRUE)
}
\arguments{
\item{FUN}{function to apply, found via \code{\link[base]{match.fun}}}

\item{FUN.VALUE}{a (generalized) vector; a template for the return value from FUN. See \code{\link[base]{vapply}}}

\item{...}{arguments to vectorize over (vectors or lists of strictly positive length, or all of zero length). See \code{\link[base]{mapply}}}

\item{MoreArgs}{a list of other arguments to FUN. Note that this is always evaluated once, even if all arguments in ... are length-zero.
Use of an anonymous function may be safer.}

\item{USE.NAMES}{logical; use names if the first ... argument has names, or if it is a character vector, use that character vector as the names.
Since this argument follows ... its name cannot be abbreviated.}
}
\description{
As \code{\link[base]{mapply}} is a multivariate version of \code{\link[base]{sapply}}, this is a multivariate version of
\code{\link[base]{vapply}}.\cr
FUN is applied to the first elements in ..., the second, and so on. Arguments are recycled if necessary.
}
