% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-wait.R
\name{waitForKey}
\alias{waitForKey}
\title{Wait for set time, or continue on input.}
\usage{
waitForKey(message = "Continuing in {n} seconds, or press any key.",
  time = 10, counter = 0.5, precision = 0.01)
}
\arguments{
\item{message}{Message to be shown during countdown. Use {n} as a placeholder for the number of seconds left.}

\item{time}{Time to wait, in seconds (approximate)}

\item{counter}{Interval to update the message, in seconds. 0 to remain static. Must otherwise be a multiple of precision to work reliably.}

\item{precision}{Polling interval to use when checking for a keypress.}
}
\value{
(invisibly) Either the string 'key' or 'timer', signifying what caused the return
}
\description{
Waits until a timer has expired, or the user has provided some input (pressed RETURN)\cr
A message is shown, with a countdown-timer if desired.
}
\note{
Unfortunately, does not work on all platforms/in all environments.\cr
It is dependent on being able to read from file('stdin'), which is different from stdin(), see \code{\link[base]{file}} for details.
In particular, it does not work in RStudio for MacOS, or Rgui for Windows or MacOS.
It does confirmed work on R run from terminal on MacOS 10.13.6\cr
Problems manifest by file('stdin') not being connected to anything, i.e. no input is received, so this function always returns \code{character(0)},
and any input typed is interpreted as a new command. To test, try with a small value of time first.
}
\seealso{
\code{\link{readline_time}} for a function that also collects your input.
}
