% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bugfix_write_table.R
\name{write.table}
\alias{write.table}
\title{Alternative for write.table}
\usage{
write.table(x, file = "", append = FALSE, quote = TRUE, sep = " ",
  eol = "\\n", na = "NA", dec = ".", row.names = TRUE,
  col.names = TRUE, qmethod = c("escape", "double"), fileEncoding = "")
}
\arguments{
\item{x, file, append, quote, sep, eol, na, dec, row.names, col.names, qmethod, fileEncoding}{See \code{\link[utils]{write.table}}}
}
\description{
Bugfix for the standard utils::write.table-function, copied from the standard-function and adapted.
When using nested data.frames and quote=TRUE, the original function first looks
at which "columns" are character of factor, and marks these for quoting, then flattens the structure to a matrix.
This functions also inspects columns that are data.frames or comparable. See the example for the difference.
}
\examples{
df <- data.frame(a='One;Two;Three',
                 b=I(data.frame(c="OtherVal",
                 d='Four;Five;Six',
                 e=4)))
write.table(df, "~/Desktop/Tempfile.csv", quote = TRUE, col.names = NA,
            sep = ";", dec = ",", qmethod = "double")
# This fails for utils::write.table, in that case Four;Five;Six is unquoted, rendering the csv useless

}
