% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{out}
\alias{out}
\title{Customary function for printing output and simalteneously writing to a logfile}
\usage{
out(..., logpath = getOption("StandardPaths")[["TextOutput"]])
}
\description{
Character vectors are printed using cat instead of print.\cr
Multiple arguments are accepted, as seperate calls\cr
Any output is first evaluated (in its entirety), then printed to files, finally to console
}
