% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devel.R
\name{stop}
\alias{stop}
\title{Customized stop, gives lines-numbers if called from script}
\usage{
stop(...)
}
\arguments{
\item{...}{arguments passed on to base::stop()}
}
\description{
Same as base::stop(), but when called from a sourced script, it also output the filename and linenumber
}
