% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fixes-extensions.R
\name{[<-.data.frame}
\alias{[<-.data.frame}
\title{Bugfix for normal data.frame assignment}
\usage{
\method{[}{data.frame}(x, i, j) <- value
}
\arguments{
\item{x}{The data.frame into which to assign}

\item{i, j}{Indices for row(s) and column(s)}

\item{value}{The new values for replacement}
}
\description{
See also \href{https://bugs.r-project.org/bugzilla3/show_bug.cgi?id=17504}{issue #17504 at R-bugzilla}
Can be used as a stand-in for the normal data.frame-assignment-method, but works around a bug:
When assigning multiple columns and specifying rows, now there is no error thrown.
Based on base::`[<-.data.frame` from R-devel, revision 75615 (so also fixes bug 17483)
}
\seealso{
\code{\link[base]{[<-.data.frame}}
}
